refactor content listing page so that it renders custom components by the config from index
